+++ src/Illuminate/Broadcasting/BroadcastController.php
@@ -13,26 +13,10 @@ class BroadcastController extends Controller
      * Authenticate the request for channel access.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return void
+     * @return \Illuminate\Http\Response
      */
     public function authenticate(Request $request)
     {
-        if (Str::startsWith($request->channel_name, ['private-', 'presence-']) &&
-            ! $request->user()) {
-            abort(403);
-        }
-
         return Broadcast::check($request);
-    }
-
-    /**
-     * Store the socket ID for the current user.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return void
-     */
-    public function rememberSocket(Request $request)
-    {
-        return Broadcast::rememberSocket($request);
     }
 }
