+++ src/Illuminate/Contracts/Broadcasting/Broadcaster.php
@@ -5,6 +5,23 @@ namespace Illuminate\Contracts\Broadcasting;
 interface Broadcaster
 {
     /**
+     * Authenticate the incoming request for a given channel.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return mixed
+     */
+    public function check($request);
+
+    /**
+     * Return the valid authentication response.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  mixed  $result
+     * @return mixed
+     */
+    public function validAuthenticationResponse($request, $result);
+
+    /**
      * Broadcast the given event.
      *
      * @param  array  $channels
